requirements.txt
substrate-interface>=1.1.0
python-dotenv>=1.0.0
